requirements.txt
# Dependências existentes
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# Novas dependências para funcionalidades avançadas
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0

# APIs Externas
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==1.3.7
twilio==8.10.0

# IA e Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2

# Gráficos e Analytics
plotly==5.17.0
matplotlib==3.8.2

# Pagamentos
stripe==7.8.0
mercadopago==2.2.0

# Cache e Performance
redis==5.0.1
celery==5.3.4

# Segurança e Validação
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0

# Telemedicina
opentok==2.9.0

# Utilitários
python-dateutil==2.8.2
Pillow==10.1.0
qrcode==7.4.2

# Geração de Relatórios
reportlab==4.0.4
